// backend/src/app/interfaces/IBook.ts
export interface IBook {
  _id?: string;
  title: string;
  author: string;
  genre:
    | "FICTION"
    | "NON_FICTION"
    | "SCIENCE"
    | "HISTORY"
    | "BIOGRAPHY"
    | "FANTASY";
  isbn: string;
  description: string;
  copies: number;
  available: boolean;
  updateAvailability: (quantity: number) => void;

  createdAt?: Date;
  updatedAt?: Date;
}